Union del punto apuestas con el resto del trabajo
