--- usuarios.csv
@@ -1,3 +1,0 @@
-id_usuario,nombre,contraseña,cantidad_apostada,fecha_última_apuesta,dinero_disponible
-pruba@mail.com,prueba,prueba,0,,1000
-fulano@mail.com,fulano,fulano ,1,2023/06/17,500
